+++ src/utils.rs
@@ -1,3 +1,5 @@
+//! Utilities to help working with kdbx-rs
+
 use std::io;
 use uuid::Uuid;
 
@@ -73,4 +75,16 @@ pub(crate) fn to_hex_string(data: &[u8]) -> String {
     }
 
     output
+}
+
+/// No-op stream cipher that does no encryption or decryption
+pub struct NullStreamCipher;
+
+impl stream_cipher::StreamCipher for NullStreamCipher {
+    fn encrypt(&mut self, _data: &mut [u8]) {
+        
+    }
+    fn decrypt(&mut self, _data: &mut [u8]) {
+    }
+    
 }